backend/requirements.txt
flask
requests
flask_cors
pyjwt
python-dotenv
pyairtable